SBRFACCTAX-8196 Неизвестная ошибка при сохранении формы "Настройки подразделений"

+++ gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/module/departmentconfig/server/SaveDepartmentCombinedHandler.java
@@ -145,10 +145,10 @@ public class SaveDepartmentCombinedHandler extends AbstractActionHandler<SaveDep
 
             // Налог на прибыль
             if (action.getTaxType() == TaxType.INCOME) {
-                paramsMap.put(DepartmentParamAliases.SUM_TAX.name(), new RefBookValue(RefBookAttributeType.NUMBER, depCombined.getSumTax()));
-                paramsMap.put(DepartmentParamAliases.SUM_DIVIDENDS.name(), new RefBookValue(RefBookAttributeType.NUMBER, depCombined.getSumDividends()));
+                paramsMap.put(DepartmentParamAliases.SUM_TAX.name(), new RefBookValue(RefBookAttributeType.NUMBER, BigDecimal.valueOf(depCombined.getSumTax())));
+                paramsMap.put(DepartmentParamAliases.SUM_DIVIDENDS.name(), new RefBookValue(RefBookAttributeType.NUMBER, BigDecimal.valueOf(depCombined.getSumDividends())));
                 paramsMap.put(DepartmentParamAliases.OBLIGATION.name(), new RefBookValue(RefBookAttributeType.REFERENCE, getFirstLong(depCombined.getObligation())));
-                paramsMap.put(DepartmentParamAliases.TAX_RATE.name(), new RefBookValue(RefBookAttributeType.NUMBER, depCombined.getTaxRate()));
+                paramsMap.put(DepartmentParamAliases.TAX_RATE.name(), new RefBookValue(RefBookAttributeType.NUMBER, BigDecimal.valueOf(depCombined.getTaxRate())));
                 paramsMap.put(DepartmentParamAliases.TYPE.name(), new RefBookValue(RefBookAttributeType.REFERENCE, getFirstLong(depCombined.getType())));
             }
